+++ src/main/kotlin/io/github/jhspetersson/turtlecommander/vfs/ZipVirtualFileSystem.kt
@@ -51,6 +51,8 @@ class ZipFileSystemProvider : VirtualFileSystemProvider {
             "whl", "egg",
             // Google Earth KML (zipped KML + assets)
             "kmz",
+            // Universal Scene Description (Pixar USD / Apple AR Quick Look, uncompressed ZIP)
+            "usdz",
         )
     }